Rename cfn__mitm module to cfn__resource_injection
