--- pacu/modules/cfn__mitm/tests/test_main.py
@@ -1,6 +1,0 @@
-from chalice.test import Client
-from pacu.modules.cfn__mitm.cfn__mitm_lambda.app import app
-
-
-def test_get_bucket_name():
-    pass
