+++ pacu/modules/cfn__resource_injection/cfn__resource_injection_lambda/.chalice/config.json
