+++ pacu/modules/cfn__resource_injection/tests/test_main.py
@@ -1,0 +1,6 @@
+from chalice.test import Client
+from pacu.modules.cfn__resource_injection.cfn__resource_injection_lambda.app import app
+
+
+def test_get_bucket_name():
+    pass
